Fix Doctrine subscriber deprecations
