+++ src/Bundle/EventListener/ORMRepositoryClassSubscriber.php
@@ -13,12 +13,13 @@ declare(strict_types=1);
 
 namespace Sylius\Bundle\ResourceBundle\EventListener;
 
+use Doctrine\Common\EventSubscriber;
 use Doctrine\ORM\EntityRepository;
 use Doctrine\ORM\Event\LoadClassMetadataEventArgs;
 use Doctrine\ORM\Events;
 use Doctrine\ORM\Mapping\ClassMetadata;
 
-final class ORMRepositoryClassSubscriber extends AbstractDoctrineSubscriber
+final class ORMRepositoryClassSubscriber extends AbstractDoctrineListener implements EventSubscriber
 {
     public function getSubscribedEvents(): array
     {
